添加了性别，用户编号,生日，个性签名的生成

+++ src/main/java/com/sixsixsix516/fakeuser/model/UserBase.java
@@ -25,7 +25,7 @@ public class UserBase {
     /**
      * 用户编号
      */
-    private Integer userNo;
+    private Long userNo;
 
     /**
      * 生日
